refactor: move parseJsonBytesAsStream to bytes module
